requirements.txt
python-telegram-bot
openai
python-dotenv
APScheduler
pytz